% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map of fatal injuries}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The input state number either as string or number}

\item{year}{The input year as an integer}
}
\value{
Draws a map plot of the locations of all the fatal injuries in the
input year and state. If the input state number is not correct it returns an
error. If the are no accidents it returns a message.
}
\description{
\code{fars_map_state} draws a state map of the location of ball the fatal
injuries of a given year and state using the \link[maps]{map} function
}
\examples{
\dontrun{fars_map_state(1,2013)}
\dontrun{fars_map_state("1","2018")}
}
