% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CourseraPackage.R
\name{fars_read}
\alias{fars_read}
\title{fars_read reads a csv file into a tibble.  This function is not exported}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{takes the filename to be read.  Include the full path of the file}
}
\value{
this function returns a tibble with the file read
}
\description{
fars_read reads a csv file into a tibble.  This function is not exported
}

