% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CourseraPackage.R
\name{fars_read_years}
\alias{fars_read_years}
\title{this function is not exported either}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{select years to be plotted}
}
\value{
a list with month year
}
\description{
this function is not exported either
}

