% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Number of fatal injuries per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{The input years either as list of strings or a as a single
string for one year}
}
\value{
Returns a \link{tibble} with all the number of fatal injuries for each
month of the years you used as input. If there is no correspnding Report the
functions returns a warning.
}
\description{
\code{fars_summarize_years} summarizes the number of fatal injuries for each
month of the years you put in.
}
\examples{
\dontrun{fars_read_years(2015:2018)}
\dontrun{fars_read_years("2015:2018")}
}
