% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CourseraPackage.R
\name{make_filename}
\alias{make_filename}
\title{this function creates the file name.  The filename is "accident_year.csv.bz2
this function is not exported}
\usage{
make_filename(year)
}
\arguments{
\item{year}{this is for the year of accidents}
}
\value{
returns the file name.
}
\description{
this function creates the file name.  The filename is "accident_year.csv.bz2
this function is not exported
}

