% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Will plot accidents as points on a map of the state by Longitude and Latitude}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The unique integer value that identifies a state}

\item{year}{An integer value that represents a year}
}
\value{
a plot of accident locations on a map of the state in which the accident
occurred
}
\description{
The location of the accidents are captured by Longitude and Latitude fields.
These fields will be used to plot the accident locations on a map of the
state in which they occur. Errors will be thrown if the inputted \code{state.num}
or \code{year} is not valid
}
\examples{
\dontrun{}
fars_map_state(40, 2013)
}
