% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Creates and returns a data frame from a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character string of the name to the file}
}
\value{
a data.frame object
}
\description{
Will throw an error if no file matches the filename
}
\examples{
fars_read(system.file("/extdata/accident_2015.csv.bz2",package="CourseraPackage"))
}
