% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read Fatality Analysis Report}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Either a path to a file, a conncetion or the data containing
the FAR. Can be all file formats which are readble by \code{\link{readr:
:read_csv}}.}
}
\value{
The function returns the data as a \link{tibble}, a data frame
 with class \link{tibble}
}
\description{
\code{fars_read} reads the annual Fatality Analysis Report of fatal injuries
suffered in motor vehicle traffic crashes using \code{\link{readr::read_csv}}
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}
\dontrun{fars_read("accident_2013.csv")}
}
