% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Will process a group of csv files in the directory}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{either a vector or list of integers representing years}
}
\value{
a list of data.frame objects read in. List may contain NULL if
a file with an input year does not exist
}
\description{
files must be in current working directory since the name make_filename
function does not generate paths but rather only the name of file
}
\examples{
fars_read_years(c(2016,2017,2018))
fars_read_years(c(1996:2017))
}
