% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Return data for given year or sequence of years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{The input years either as list of strings or a as
a single string for one year}
}
\value{
Returns a list of \link{tibble}s with all the data for the years you
 used as input. If there is no correspnding report the functions returns a
 warning.
}
\description{
\code{fars_read_years} searches your working directory for files for the
corresponding Fatality Analysis Reports for the years you input and returns a
combined table with all the data for the years you input
}
\examples{
\dontrun{fars_read_years(c("2014","2015"))}
\dontrun{fars_read_years("2015")}
}
