% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Creates one summary output from all CSV files with the inputed years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{either a vector or list of integers representing years}
}
\value{
a matrix where columns represent year and rows represent months
and the value of each cell is the total number of occurrences
}
\description{
Will first merge all csv files and then summarise the resulting single data frame.
}
\examples{
fars_summarize_years(c(2016, 2017, 2018))
fars_summarize_years(c(1996:2017))
}
