% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Returns a string that can be a name to a CSV file, where the input \code{year} will
be part of the string}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a number representing a year}
}
\value{
a string of the pattern: "accident_{year}.csv.bz2"
}
\description{
Function will throw an error if \code{year} is not an integer
}
\examples{
make_filename(2018)
}
