% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Make .csv.bz2 file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The input year as integer}
}
\value{
Returns a file name as string according to this patter:
"accident_year.csv.bz2"
}
\description{
\code{make_filename} takes a year and creates a filename string according to
this year
}
\examples{
\dontrun{make_filename("2015")}
}
